+++ stellar_sdk/soroban_server_async.py
@@ -53,7 +53,7 @@ class SorobanServerAsync:
     async def get_health(self) -> GetHealthResponse:
         """General node health check.
 
-        See `Soroban Documentation - getHealth <https://soroban.stellar.org/api/methods/getHealth>`_
+        See `Soroban RPC Documentation - getHealth <https://developers.stellar.org/docs/data/rpc/api-reference/methods/getHealth>`_
 
         :return: A :class:`GetHealthResponse <stellar_sdk.soroban_rpc.GetHealthResponse>` object.
         :raises: :exc:`SorobanRpcErrorResponse <stellar_sdk.exceptions.SorobanRpcErrorResponse>` - If the Soroban-RPC instance returns an error response.
@@ -74,7 +74,7 @@ class SorobanServerAsync:
     ) -> GetEventsResponse:
         """Fetch a list of events that occurred in the ledger range.
 
-        See `Soroban Documentation - getEvents <https://soroban.stellar.org/api/methods/getEvents>`_
+        See `Soroban RPC Documentation - getEvents <https://developers.stellar.org/docs/data/rpc/api-reference/methods/getEvents>`_
 
         :param start_ledger: The first ledger to include in the results.
         :param filters: A list of filters to apply to the results.
@@ -97,6 +97,8 @@ class SorobanServerAsync:
     async def get_network(self) -> GetNetworkResponse:
         """General info about the currently configured network.
 
+        See `Soroban RPC Documentation - getNetwork <https://developers.stellar.org/docs/data/rpc/api-reference/methods/getNetwork>`_
+
         :return: A :class:`GetNetworkResponse <stellar_sdk.soroban_rpc.GetNetworkResponse>` object.
         :raises: :exc:`SorobanRpcErrorResponse <stellar_sdk.exceptions.SorobanRpcErrorResponse>` - If the Soroban-RPC instance returns an error response.
         """
@@ -109,6 +111,8 @@ class SorobanServerAsync:
 
     async def get_latest_ledger(self) -> GetLatestLedgerResponse:
         """Fetches the latest ledger meta info from network which Soroban-RPC is connected to.
+
+        See `Soroban RPC Documentation - getLatestLedger <https://developers.stellar.org/docs/data/rpc/api-reference/methods/getLatestLedger>`_
 
         :return: A :class:`GetLatestLedgerResponse <stellar_sdk.soroban_rpc.GetLatestLedgerResponse>` object.
         :raises: :exc:`SorobanRpcErrorResponse <stellar_sdk.exceptions.SorobanRpcErrorResponse>` - If the Soroban-RPC instance returns an error response.
@@ -129,7 +133,7 @@ class SorobanServerAsync:
         or any other ledger entry. This is a backup way to access your contract data
         which may not be available via events or simulateTransaction.
 
-        See `Soroban Documentation - getLedgerEntries <https://soroban.stellar.org/api/methods/getLedgerEntries>`_
+        See `Soroban RPC Documentation - getLedgerEntries <https://developers.stellar.org/docs/data/rpc/api-reference/methods/getLedgerEntries>`_
 
         :param keys: The ledger keys to fetch.
         :return: A :class:`GetLedgerEntriesResponse <stellar_sdk.soroban_rpc.GetLedgerEntryResponse>` object.
@@ -145,7 +149,7 @@ class SorobanServerAsync:
     async def get_transaction(self, transaction_hash: str) -> GetTransactionResponse:
         """Fetch the specified transaction.
 
-        See `Soroban Documentation - getTransaction <https://soroban.stellar.org/api/methods/getTransaction>`_
+        See `Soroban RPC Documentation - getTransaction <https://developers.stellar.org/docs/data/rpc/api-reference/methods/getTransaction>`_
 
         :param transaction_hash: The hash of the transaction to fetch.
         :return: A :class:`GetTransactionResponse <stellar_sdk.soroban_rpc.GetTransactionResponse>` object.
@@ -165,7 +169,7 @@ class SorobanServerAsync:
     ) -> SimulateTransactionResponse:
         """Submit a trial contract invocation to get back return values, expected ledger footprint, and expected costs.
 
-        See `Soroban Documentation - simulateTransaction <https://soroban.stellar.org/api/methods/simulateTransaction>`_
+        See `Soroban RPC Documentation - simulateTransaction <https://developers.stellar.org/docs/data/rpc/api-reference/methods/simulateTransaction>`_
 
         :param transaction_envelope: The transaction to simulate. It should include exactly one operation,
             which must be one of :class:`RestoreFootprint <stellar_sdk.operation.RestoreFootprintOperation>`,
@@ -203,7 +207,7 @@ class SorobanServerAsync:
     ) -> SendTransactionResponse:
         """Submit a real transaction to the Stellar network. This is the only way to make changes "on-chain".
 
-        See `Soroban Documentation - sendTransaction <https://soroban.stellar.org/api/methods/sendTransaction>`_
+        See `Soroban RPC Documentation - sendTransaction <https://developers.stellar.org/docs/data/rpc/api-reference/methods/sendTransaction>`_
 
         :param transaction_envelope: The transaction to send.
         :return: A :class:`SendTransactionResponse <stellar_sdk.soroban_rpc.SendTransactionResponse>` object.
@@ -220,6 +224,48 @@ class SorobanServerAsync:
             params=SendTransactionRequest(transaction=xdr),
         )
         return await self._post(request, SendTransactionResponse)
+
+    async def get_fee_stats(self) -> GetFeeStatsResponse:
+        """General info about the fee stats.
+
+        See `Soroban RPC Documentation - getFeeStats <https://developers.stellar.org/docs/data/rpc/api-reference/methods/getFeeStats>`_
+
+        :return: A :class:`GetFeeStatsResponse <stellar_sdk.soroban_rpc.GetFeeStatsResponse>` object.
+        :raises: :exc:`SorobanRpcErrorResponse <stellar_sdk.exceptions.SorobanRpcErrorResponse>` - If the Soroban-RPC instance returns an error response.
+        """
+        request: Request = Request(
+            id=_generate_unique_request_id(),
+            method="getFeeStats",
+            params=None,
+        )
+        return await self._post(request, GetFeeStatsResponse)
+
+    async def get_transactions(
+        self,
+        start_ledger: int,
+        cursor: str = None,
+        limit: int = None,
+    ) -> GetTransactionsResponse:
+        """Fetch a detailed list of transactions starting from the user specified starting point that you can paginate
+        as long as the pages fall within the history retention of their corresponding RPC provider.
+
+        See `Soroban RPC Documentation - getTransactions <https://developers.stellar.org/docs/data/rpc/api-reference/methods/getTransactions>`_
+
+        :param start_ledger: The first ledger to include in the results.
+        :param cursor: A cursor value for use in pagination.
+        :param limit: The maximum number of records to return.
+        :return: A :class:`GetTransactionsResponse <stellar_sdk.soroban_rpc.GetTransactionsResponse>` object.
+        :raises: :exc:`SorobanRpcErrorResponse <stellar_sdk.exceptions.SorobanRpcErrorResponse>` - If the Soroban-RPC instance returns an error response.
+        """
+        pagination = PaginationOptions(cursor=cursor, limit=limit)
+        data = GetTransactionsRequest(
+            startLedger=str(start_ledger),
+            pagination=pagination,
+        )
+        request: Request = Request[GetTransactionsRequest](
+            id=_generate_unique_request_id(), method="getTransactions", params=data
+        )
+        return await self._post(request, GetTransactionsResponse)
 
     async def load_account(self, account_id: str) -> Account:
         """Load an account from the server, you can use the returned account